Use empty string as default value for reporter type
### What's done:
* Extract `createReporterFlag` method
* Add test
